requirements.txt
# Base requirements
python-dotenv==0.19.0
requests==2.31.0

# PDF and document processing
PyPDF2==3.0.1
python-docx==0.8.11

# Optional: NLP capabilities
# Uncomment the following lines if you need NLP features
# spacy>=3.7.0
# en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0-py3-none-any.whl

# LLM providers
openai==1.3.7
anthropic>=0.8.1
ollama>=0.1.5  # Python client for Ollama

# Optional: Web interface
# Uncomment the following lines if you want to use the web interface
# fastapi>=0.104.1
# uvicorn[standard]>=0.23.2
# jinja2>=3.1.2
# python-multipart>=0.0.6

# Ollama integration
# ollama==0.1.5  # Removed as it refers to a different package

# Web app
flask==2.0.1
flask-cors==3.0.10
langchain==0.0.267
langchain-community>=0.0.6
pydantic==1.9.0
markdown==3.4.1

# Testing
pytest==7.3.1
